requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
omegaconf>=2.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
plotly>=5.15.0

# PyTorch ecosystem
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
einops>=0.6.0
albumentations>=1.3.0
timm>=0.9.0

# Training and optimization
pytorch-lightning>=2.0.0
accelerate>=0.20.0
transformers>=4.30.0

# Evaluation and metrics
torchmetrics>=1.0.0
lpips>=0.1.4
clean-fid>=0.1.35
scikit-learn>=1.3.0

# Visualization and demo
streamlit>=1.25.0
gradio>=3.40.0
wandb>=0.15.0
tensorboard>=2.13.0

# Development tools
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0

# Optional: for advanced features
opencv-python>=4.8.0
pillow>=10.0.0
scipy>=1.11.0
